fixed bug
If the gate toggle distance is nonzero, then e.g. the player can walk past it horizontally (above/below the gate so they don't activate it), then go to the same y-level as the gate and go back the other way. If they didn't go far enough to change the `playerSide`, then the gate won't toggle as they go through.