changed traversability storage format
Now it stores it as a list of pairs of states `(x, y)` such that you can go from state `x` to state `y`. This is a necessary change because the paths are not reversible: if you can go from state `x` to state `y`, you might not be able to go from state `y` to state `x` because a gate might have closed behind you.