changed definition of connectivity
Now it is the probability that if you choose two random exits A and B and enter with the doors either toggled or not, the player can go from A to B.

(Under this definition, most existing rooms have a connectivity of 1 regardless of where they spawn. But this will change because I plan on adding many rooms that can only be traversed in one direction or can only be traversed if the gates are in a certain state).