fixed bug: lizards would sometimes disappear
This would happen if the lizard was completely horizontal/vertical and perfectly lined up with the boundary of an entity chunk. Since lizard hitboxes previously had a width/height of zero in this case, this would result in it not being considered inside of any entity chunk, and so it would appear as if the lizard had just ceased to exist.

Also fixed the much-more-obvious bug that entities should not be able to push themselves.