reverted to previous tower-style generation algorithm
Reasoning: while the open world generation is cool, it led to the game feeling very aimless. I considered adding secrets and stuff to give it a sense of purpose, but the fact remained that there was no clear way to measure progression or end the game (apart from just counting "number of rooms explored", which is not very exciting for the player). On the other hand, the tower has a very clear progression (the player goes up), so I changed it back to the tower.