fixed some inescapable dead ends
The problem was that the room "partial-control-room" could generate with only an upwards exit, making it impossible to leave. This was not detected by the level generation algorithm, because although the level was impossible to complete from that room, there was no gate state from which it was impossible to complete.